requirements.txt
pyyaml
praw
discord.py
